--- src/c-createPost/services/createPost.service.js
@@ -1,12 +1,0 @@
-const axios = require('axios');
-
-require('dotenv').config();
-const {
-    API_KEY
-} = process.env;
-
-// agregamos los servicios que queremos exportar
-module.exports = {
-
-
-}
